Move get_subscription_id_list into Core
